% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_readme_vignetts.R
\name{add_readme_to_vignette}
\alias{add_readme_to_vignette}
\title{Add README to vignette}
\usage{
add_readme_to_vignette(source, vignette_title = NULL, vignette_slug = "README")
}
\arguments{
\item{source}{String. The directory path where the package is deployed.}

\item{vignette_title}{String. The intended title of the vignette. If not provided, no title will be attached.}

\item{vignette_slug}{String. filename to be used as the vignette. By default this will be README.}
}
\description{
If a package does not have a vignette but has a \code{README.md}, make the contents of the \code{README.md} be the vignette
}
\details{
Review the DISCRIPTION file and file structure and add the necessary dependencies and files.
}
