% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme2vignette-package.R
\docType{package}
\name{readme2vignette-package}
\alias{readme2vignette}
\alias{readme2vignette-package}
\title{readme2vignette: Convert README to Vignette during Package installation}
\description{
Convert `README.md` to vignettes when installing packages without vignettes.
}
\author{
\strong{Maintainer}: Mao Kobayashi \email{kobamao.jp@gmail.com}

}
\keyword{internal}
